+++ ledger/tests/test_api/test_char_journal_api.py
@@ -2,6 +2,7 @@ from typing import Optional
 
 from ninja import NinjaAPI
 
+from django.db.models import Q
 from django.test import TestCase
 
 from app_utils.testing import create_user_from_evecharacter
@@ -116,7 +117,7 @@ class ManageApiJournalCharEndpointsTest(TestCase):
 
     def test_get_character_journal_api_single_ref_types(self):
         self.client.force_login(self.user)
-        url = "/ledger/api/account/1001/wallet/?type_refs=bounty_prizes,player_donation,contract_reward,transaction_tax,industry_job_tax,market_escrow"
+        url = "/ledger/api/account/1001/wallet/?type_refs=bounty_prizes,player_donation,contract_reward,transaction_tax,industry_job_tax,market_escrow,insurance"
 
         journal_query = CharacterWalletJournalEntry.objects.filter(
             character__character__character_id=1001